refactor directory subscribe
